pixel-test-python/requirements.txt
request==2.32.3
beautifulsoup4==4.12.3
